[CI skip] A lot of refactoring
