+++ src/main/java/me/mrCookieSlime/Slimefun/Lists/RecipeType.java
@@ -9,6 +9,8 @@ import java.util.Locale;
 import java.util.Set;
 import java.util.function.BiConsumer;
 
+import javax.annotation.ParametersAreNonnullByDefault;
+
 import org.bukkit.ChatColor;
 import org.bukkit.Keyed;
 import org.bukkit.Material;
@@ -138,16 +140,32 @@ public class RecipeType implements Keyed {
     }
 
     @Override
-    public NamespacedKey getKey() {
+    public final NamespacedKey getKey() {
         return key;
     }
 
+    @Override
+    public final boolean equals(Object obj) {
+        if (obj instanceof RecipeType) {
+            return ((RecipeType) obj).getKey().equals(this.getKey());
+        } else {
+            return false;
+        }
+    }
+
+    @Override
+    public final int hashCode() {
+        return getKey().hashCode();
+    }
+
+    @ParametersAreNonnullByDefault
     private static void registerBarterDrop(ItemStack[] recipe, ItemStack output) {
         if (SlimefunPlugin.getMinecraftVersion().isAtLeast(MinecraftVersion.MINECRAFT_1_16)) {
             SlimefunPlugin.getRegistry().getBarteringDrops().add(output);
         }
     }
 
+    @ParametersAreNonnullByDefault
     private static void registerMobDrop(ItemStack[] recipe, ItemStack output) {
         String mob = ChatColor.stripColor(recipe[4].getItemMeta().getDisplayName()).toUpperCase(Locale.ROOT).replace(' ', '_');
         EntityType entity = EntityType.valueOf(mob);
